Simplify token case normalization and handling
Centralize lowercase normalization of the label and mnemonic in only one place of the parser. Delegate case handling to code that processes operands elsewhere.

Remove case dependency from tests. More case handling work is needed to accept assembler source in any combination of uppercase and lowercase, so it's early for these tests.