Implement -s option, fix -o argument bug, tweak output
Add to the assembler the -s option for saving the symbol table to a file in .sym CP/M format.

Fix a bug that ignores the -o argument when building the output file name and - is supplied as the input file.

Add the missing extension of the default output file to the help message printed when running the assembler with the -h option.

Revise the formatting of verbose output.